+++ app/src/main/java/com/example/installhook/MainActivity.java
@@ -8,7 +8,7 @@ import de.robv.android.xposed.XposedBridge;
 import de.robv.android.xposed.XposedHelpers;
 import de.robv.android.xposed.callbacks.XC_LoadPackage;
 
-public class MainActivity extends AppCompatActivity implements IXposedHookLoadPackage {
+public class MainActivity extends AppCompatActivity {
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
@@ -16,33 +16,5 @@ public class MainActivity extends AppCompatActivity implements IXposedHookLoadPa
         setContentView(R.layout.activity_main);
     }
 
-    @Override
-    public void handleLoadPackage(XC_LoadPackage.LoadPackageParam lpparam) throws Throwable {
-        XposedBridge.log(lpparam.packageName);
-        if (!"com.android.packageinstaller".equals(lpparam.packageName)) {
-            return;
-        }
 
-        // 自动开始安装
-        XposedHelpers.findAndHookMethod("com.android.packageinstaller.PackageInstallerActivity", lpparam.classLoader, "bindUi", new XC_MethodHook() {
-            @Override
-            protected void afterHookedMethod(MethodHookParam param) throws Throwable {
-                Object mAlert = XposedHelpers.getObjectField(param.thisObject, "mAlert");
-                Object button = XposedHelpers.callMethod(mAlert, "getButton", -1);
-                XposedHelpers.callMethod(button,"callOnClick");
-                super.afterHookedMethod(param);
-            }
-        });
-
-        // 安装成功自动点击完成
-        XposedHelpers.findAndHookMethod("com.android.packageinstaller.InstallSuccess", lpparam.classLoader, "onCreate", new XC_MethodHook() {
-            @Override
-            protected void afterHookedMethod(MethodHookParam param) throws Throwable {
-                Object mAlert = XposedHelpers.getObjectField(param.thisObject, "mAlert");
-                Object button = XposedHelpers.callMethod(mAlert, "getButton", -1);
-                XposedHelpers.callMethod(button,"callOnClick");
-                super.afterHookedMethod(param);
-            }
-        });
-    }
 }
